requirements.txt
Flask-Bcrypt==0.7.1
Flask-MySQLdb==1.0.1
Flask-SocketIO>=5.0.0
Flask>=2.0.0
eventlet==0.33.3
python-dotenv==0.19.2
python-engineio==4.5.1
python-socketio==5.8.0
